requirements.txt
# Database
neo4j>=5.9.0

# Backend & API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
asyncio>=3.4.3
starlette>=0.30.0

# AI
openai>=1.0.0
langchain>=0.0.300
python-dotenv>=1.0.0

# Frontend
customtkinter>=5.2.0
pillow>=10.0.0
tkinter-tooltip>=0.2.0  # If not available, can be replaced

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Utilities
pyyaml>=6.0
python-multipart>=0.0.6
tqdm>=4.65.0
